requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.10
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==1.7.1
python3-openid==3.2.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
simplejwt==2.0.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
